Started reworking multiplication for consistency checkers.
