+++ pumpkin-crates/propagators/src/propagators/arithmetic/mod.rs
@@ -2,15 +2,15 @@
 pub(crate) mod absolute_value;
 pub(crate) mod binary;
 pub(crate) mod integer_division;
-pub(crate) mod integer_multiplication;
 pub(crate) mod linear_less_or_equal;
 pub(crate) mod linear_not_equal;
 pub(crate) mod maximum;
+mod multiplication;
 
 pub use absolute_value::*;
 pub use binary::*;
 pub use integer_division::*;
-pub use integer_multiplication::*;
 pub use linear_less_or_equal::*;
 pub use linear_not_equal::*;
 pub use maximum::*;
+pub use multiplication::*;
